+++ Content/Tiles/Pedestal.cs
@@ -133,7 +133,7 @@ namespace Radiance.Content.Tiles
                             radianceAdjustingTexture,
                             position,
                             null,
-                            Color.Lerp(Radiance.RadianceColor1 * fill, Radiance.RadianceColor2 * fill, fill * (float)RadianceUtils.SineTiming(5)),
+                            Color.Lerp(Radiance.RadianceColor1 * fill, Radiance.RadianceColor2 * fill, fill * RadianceUtils.SineTiming(5)),
                             0,
                             origin,
                             1,
@@ -153,7 +153,7 @@ namespace Radiance.Content.Tiles
                          0.65f * fill * strength,
                          0.5f * fill * strength
                         ),
-                        fill * (float)RadianceUtils.SineTiming(20)).ToVector3());
+                        fill * RadianceUtils.SineTiming(20)).ToVector3());
                     }
 
                     if (Main.LocalPlayer.GetModPlayer<RadiancePlayer>().debugMode)
@@ -338,8 +338,8 @@ namespace Radiance.Content.Tiles
                 containerPlaced = container;
                 if (container.ContainerQuirk == ContainerQuirkEnum.Leaking) container.LeakRadiance();
                 if (container.ContainerQuirk != ContainerQuirkEnum.CantAbsorb && container.ContainerQuirk != ContainerQuirkEnum.CantAbsorbNonstandardTooltip) 
-                    container.AbsorbStars(vector + (Vector2.UnitY * 5 * (float)RadianceUtils.SineTiming(30) - yCenteringOffset / 5));
-                if(container.ContainerMode != ContainerModeEnum.InputOnly) container.FlareglassCreation(vector + (Vector2.UnitY * 5 * (float)RadianceUtils.SineTiming(30) - yCenteringOffset / 5));
+                    container.AbsorbStars(vector + (Vector2.UnitY * 5 * RadianceUtils.SineTiming(30) - yCenteringOffset / 5));
+                if(container.ContainerMode != ContainerModeEnum.InputOnly) container.FlareglassCreation(vector + (Vector2.UnitY * 5 * RadianceUtils.SineTiming(30) - yCenteringOffset / 5));
 
                 aoeCircleInfo =
                     (
@@ -367,7 +367,7 @@ namespace Radiance.Content.Tiles
                 {
                     if (Main.rand.NextBool(3))
                     {
-                        int f = Dust.NewDust(pos - new Vector2(0, -5 * (float)RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.TeleportationPotion, 0, 0);
+                        int f = Dust.NewDust(pos - new Vector2(0, -5 * RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.TeleportationPotion, 0, 0);
                         Main.dust[f].velocity *= 0.3f;
                         Main.dust[f].scale = 0.8f;
                     }
@@ -376,7 +376,7 @@ namespace Radiance.Content.Tiles
                 {
                     for (int i = 0; i < 5; i++)
                     {
-                        int f = Dust.NewDust(pos - new Vector2(0, -5 * (float)RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.TeleportationPotion, 0, 0);
+                        int f = Dust.NewDust(pos - new Vector2(0, -5 * RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.TeleportationPotion, 0, 0);
                         Main.dust[f].velocity *= 0.3f;
                         Main.dust[f].scale = Main.rand.NextFloat(1.3f, 1.7f);
                     }
@@ -394,7 +394,7 @@ namespace Radiance.Content.Tiles
                     );
                 if (Main.GameUpdateCount % 120 == 0)
                 {
-                    int f = Dust.NewDust(pos - new Vector2(0, -5 * (float)RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.PurpleCrystalShard, 0, 0);
+                    int f = Dust.NewDust(pos - new Vector2(0, -5 * RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.PurpleCrystalShard, 0, 0);
                     Main.dust[f].velocity *= 0.1f;
                     Main.dust[f].noGravity = true;
                     Main.dust[f].scale = Main.rand.NextFloat(1.2f, 1.4f);
@@ -403,7 +403,7 @@ namespace Radiance.Content.Tiles
                 {
                     for (int i = 0; i < 5; i++)
                     {
-                        int f = Dust.NewDust(pos - new Vector2(0, -5 * (float)RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.PurpleCrystalShard, 0, 0);
+                        int f = Dust.NewDust(pos - new Vector2(0, -5 * RadianceUtils.SineTiming(30) + 2) - new Vector2(8, 8), 16, 16, DustID.PurpleCrystalShard, 0, 0);
                         Main.dust[f].velocity *= 0.3f;
                         Main.dust[f].noGravity = true;
                         Main.dust[f].scale = Main.rand.NextFloat(1.3f, 1.7f);
